tests/subworkspace/pyproject.toml
[project]
name = "subworkspace"
version = "0.1.0"
description = "Workspace for testing nox-uv"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Private :: Do Not Upload",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = ["nox-uv"]

[tool.uv]
sources.nox-uv = { path = "../../", editable = true }

[tool.uv.workspace]
members = [
    "package-one",
    "package-two",
]
